Refactor in hopes of inheriting some of the analysis functions
