+++ app/bot_communities/retweet_analyzer.py
@@ -2,80 +2,17 @@
 import os
 from functools import lru_cache
 
-from pandas import DataFrame, read_csv
+from pandas import DataFrame
 import matplotlib.pyplot as plt
 import plotly.express as px
 import squarify
 
 from app import APP_ENV, seek_confirmation
-from app.decorators.datetime_decorators import dt_to_s, logstamp, dt_to_date, s_to_dt, s_to_date
+from app.decorators.datetime_decorators import logstamp
 from app.decorators.number_decorators import fmt_n
-from app.bot_communities.spectral_clustermaker import SpectralClustermaker
+from app.bot_communities.csv_storage import LocalStorage
 from app.bot_communities.tokenizers import Tokenizer
 from app.bot_communities.token_analyzer import summarize_token_frequencies, train_topic_model, parse_topics, LdaMulticore
-
-BATCH_SIZE = 50_000 # we are talking about downloading millions of records
-#ROWS_LIMIT = os.getenv("ROWS_LIMIT")
-
-class RetweetAnalyzer:
-    def __init__(self, tokenize=None):
-        self.clustermaker = SpectralClustermaker()
-        self.n_clusters = self.clustermaker.n_clusters
-        self.bq_service = self.clustermaker.grapher.bq_service
-        self.local_dirpath = self.clustermaker.local_dirpath
-
-        #self.tweets_filepath = os.path.join(self.local_dirpath, "tweets.csv")
-        self.retweets_filepath = os.path.join(self.local_dirpath, "retweets.csv")
-
-        self.retweets_dirpath = os.path.join(self.local_dirpath, "retweets")
-        if not os.path.exists(self.retweets_dirpath):
-            os.makedirs(self.retweets_dirpath)
-
-        self.tokenize = tokenize or Tokenizer().custom_stems # using the faster version, not the spacy version, due to time constraints
-
-    def load_retweets(self):
-        """
-        Loads or downloads bot community tweets to/from CSV.
-        """
-
-        if os.path.isfile(self.retweets_filepath):
-            print("READING BOT COMMUNITY RETWEETS FROM CSV...")
-            self.retweets_df = read_csv(self.retweets_filepath) # DtypeWarning: Columns (6) have mixed types.Specify dtype option on import or set low_memory=False
-            #if ROWS_LIMIT:
-            #    self.retweets_df = read_csv(local_csv_filepath, nrows=int(ROWS_LIMIT))
-            #else:
-            #    self.retweets_df = read_csv(local_csv_filepath)
-        else:
-            print("DOWNLOADING BOT COMMUNITY RETWEETS...")
-            counter = 0
-            records = []
-            for row in self.bq_service.download_n_bot_community_retweets_in_batches(self.n_clusters):
-                records.append({
-                    "community_id": row.community_id,
-                    "user_id": row.user_id,
-                    "user_screen_name_count": row.user_screen_name_count,
-                    "user_screen_names": row.user_screen_names,
-                    "user_created_at": dt_to_s(row.user_created_at),
-
-                    "retweeted_user_id": row.retweeted_user_id,
-                    "retweeted_user_screen_name": row.retweeted_user_screen_name,
-
-                    "status_id": row.status_id,
-                    "status_text": row.status_text,
-                    "status_created_at": dt_to_s(row.status_created_at)
-                })
-                counter+=1
-                if counter % BATCH_SIZE == 0: print(logstamp(), fmt_n(counter))
-
-            self.retweets_df = DataFrame(records)
-            self.retweets_df.index.name = "row_id"
-            self.retweets_df.index += 1
-            print("WRITING TO FILE...")
-            self.retweets_df.to_csv(self.retweets_filepath)
-
-    @property
-    def community_ids(self):
-        return list(self.retweets_df["community_id"].unique())
 
 
 class CommunityRetweetsAnalyzer:
@@ -201,28 +138,31 @@ class CommunityRetweetsAnalyzer:
 
 if __name__ == "__main__":
 
-    analyzer = RetweetAnalyzer()
-    analyzer.load_retweets()
-    print(analyzer.retweets_df.head())
+    storage = LocalStorage()
+    storage.load_retweets()
+    print(storage.retweets_df.head())
 
     seek_confirmation()
 
-    for community_id in analyzer.community_ids:
+    for community_id in storage.retweet_community_ids:
         community_analyzer = CommunityRetweetsAnalyzer(
             community_id=community_id,
-            community_retweets_df=analyzer.retweets_df[analyzer.retweets_df["community_id"] == community_id],
-            local_dirpath=os.path.join(analyzer.local_dirpath, f"community-{community_id}")
+            community_retweets_df=storage.retweets_df[storage.retweets_df["community_id"] == community_id],
+            local_dirpath=os.path.join(storage.local_dirpath, f"community-{community_id}")
         )
 
-        print(logstamp(), community_id, "CHARTS...")
+        print("------------------------")
+        print(logstamp(), "COMMUNITY", community_id, "CHARTS...")
         community_analyzer.generate_most_retweets_chart()
         community_analyzer.generate_most_retweeters_chart()
 
-        print(logstamp(), community_id, "TOKENS...")
+        print("------------------------")
+        print(logstamp(), "COMMUNITY", community_id, "TOKENS...")
         community_analyzer.top_tokens_df
         community_analyzer.save_top_tokens()
         community_analyzer.generate_top_tokens_wordcloud()
 
-        print(logstamp(), community_id, "TOPICS...")
+        print("------------------------")
+        print(logstamp(), "COMMUNITY", community_id, "TOPICS...")
         #community_analyzer.topics_df # TODO: taking too long
         #community_analyzer.save_topics() # TODO: taking too long
